+++ src/voxels/ChunkManager/Type.java
@@ -7,13 +7,13 @@ package voxels.ChunkManager;
 public final class Type {
     
     
-    public static final short AIR = 0;
-    public static final short WATER = 1;
-    public static final short WOOD = 2;
-    public static final short LEAVES = 3;
-    public static final short STONE = 4;
-    public static final short CLOUD = 5;
-    public static final short UNBREAKABLE = 6;
-    public static final short GRASS = 7;
-    public static final short DIRT = 8;
+    public static final byte AIR = 0;
+    public static final byte WATER = 1;
+    public static final byte WOOD = 2;
+    public static final byte LEAVES = 3;
+    public static final byte STONE = 4;
+    public static final byte CLOUD = 5;
+    public static final byte UNBREAKABLE = 6;
+    public static final byte GRASS = 7;
+    public static final byte DIRT = 8;
 }
